refactor: Form component with multiple exports replaces individual files
