+++ src/components/Form.scss
@@ -51,3 +51,49 @@
     }
   }
 }
+
+.form-group {
+  margin-bottom: 0.5rem;
+}
+
+.form-input {
+  width: 100%;
+  height: 45px;
+  margin: 0.4rem 0;
+  padding: 0.2rem 1rem 0.2rem 1rem;
+  border: 2px solid #dddddd;
+  border-radius: 3px;
+  box-sizing: border-box;
+
+  font-size: 1rem;
+
+  &:focus {
+    outline: none;
+    border-color: #aaaaaa;
+  }
+
+  &::placeholder {
+    color: #aaaaaa;
+  }
+}
+
+.form-select {
+  width: 100%;
+  height: 45px;
+	font-size: 1rem;
+  padding: 0.2rem 1rem 0.2rem 1rem;
+	max-width: 100%;
+	box-sizing: border-box;
+	margin: 0.4rem 0;
+	border: 2px solid #dddddd;
+	border-radius: 3px;
+	-moz-appearance: none;
+	-webkit-appearance: none;
+	appearance: none;
+  background-color: #fff;
+  
+  &:focus {
+    outline: none;
+    border-color: #aaaaaa;
+  }
+}
