--- src/components/FormGroup.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const FormGroup = ({ children }) => <div style={{ marginBottom: '0.5rem' }}>{children}</div>;
-
-export default FormGroup;
