--- src/components/FormInput.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import './FormInput.scss';
-
-const FormInput = (props) => (
-  <input className="form-input" {...props} />
-);
-
-export default FormInput;
