--- src/components/FormInput.scss
@@ -1,20 +1,0 @@
-.form-input {
-  width: 100%;
-  height: 45px;
-  margin: 0.4rem 0;
-  padding: 0.2rem 1rem 0.2rem 1rem;
-  border: 2px solid #dddddd;
-  border-radius: 3px;
-  box-sizing: border-box;
-
-  font-size: 1rem;
-
-  &:focus {
-    outline: none;
-    border-color: #aaaaaa;
-  }
-
-  &::placeholder {
-    color: #aaaaaa;
-  }
-}
