--- src/components/FormSelect.scss
@@ -1,20 +1,0 @@
-.form-select {
-  width: 100%;
-  height: 45px;
-	font-size: 1rem;
-  padding: 0.2rem 1rem 0.2rem 1rem;
-	max-width: 100%;
-	box-sizing: border-box;
-	margin: 0.4rem 0;
-	border: 2px solid #dddddd;
-	border-radius: 3px;
-	-moz-appearance: none;
-	-webkit-appearance: none;
-	appearance: none;
-  background-color: #fff;
-  
-  &:focus {
-    outline: none;
-    border-color: #aaaaaa;
-  }
-}
